perf: drop NormalizingFlows/OptimizationBBO/OptimizationNLopt, move SciMLSensitivity to an extension
NormalizingFlows contributed nothing: PlanarLayer and transformed come from
Bijectors, fchain from FunctionChains, and AbstractNormalizingFlow is defined
in NoLimits itself. Dropping it removes Flux, GPUArrays, MLUtils, OneHotArrays
and MonotonicSplines.

OptimizationNLopt was never referenced in src/. OptimizationBBO was used for a
single parentmodule identity check, now a module-name comparison; users passing
a BBO optimizer install it themselves.

SciMLSensitivity was imported for five recursive_*!/allocate_vjp methods on
DEParams used only by the Enzyme reverse-ODE adjoint. They move to
NoLimitsSciMLSensitivityExt, which drops Enzyme, Enzyme_jll, GPUCompiler,
ReverseDiff, Tracker and LLVM from a default install.

using NoLimits: 12.2s -> 6.9s.
