+++ Project.toml
@@ -30,11 +30,8 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 Lux = "b2108857-7c20-44ae-9111-449ecde12c47"
 MCMCChains = "c7f686f2-ff18-58e9-bc7b-31028e88f75d"
-NormalizingFlows = "50e4474d-9f12-44b7-af7a-91ab30ff6256"
 Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
-OptimizationBBO = "3e6eede4-6085-4f62-9a71-46d9bc1eb92b"
-OptimizationNLopt = "4e6fcdb7-1186-4e1f-a706-475e75c168bb"
 OptimizationOptimJL = "36348300-93cb-4f02-beb5-3c3902f8871e"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
@@ -42,7 +39,6 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
 RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
-SciMLSensitivity = "1ed8b502-d754-442c-8d5d-10ac956f44a1"
 SciMLStructures = "53ae85a6-f571-4167-b2af-e1d143709226"
 SimpleChains = "de6bee2f-e2f4-4ec7-b6ed-219cc6f6e9e5"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
@@ -55,10 +51,12 @@ Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 [weakdeps]
 Copulas = "ae264745-0b69-425e-9d9d-cf662c5eec93"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+SciMLSensitivity = "1ed8b502-d754-442c-8d5d-10ac956f44a1"
 
 [extensions]
 NoLimitsCopulasExt = "Copulas"
 NoLimitsMakieExt = "Makie"
+NoLimitsSciMLSensitivityExt = "SciMLSensitivity"
 
 [compat]
 AdvancedVI = "0.6.2, 0.7"
@@ -88,11 +86,8 @@ Logging = "1.11.0"
 Lux = "1.29.4"
 MCMCChains = "7.7.0"
 Makie = "0.24"
-NormalizingFlows = "0.2"
 Optimisers = "0.4.7"
 Optimization = "5.4.0 - 5.6"
-OptimizationBBO = "0.4.5"
-OptimizationNLopt = "0.3.11"
 OptimizationOptimJL = "0.4.9"
 OrdinaryDiffEq = "6, 7.1"
 ProgressMeter = "1.11.0"
