+++ src/estimation/pooled.jl
@@ -8,7 +8,6 @@ using SciMLBase
 using ComponentArrays
 using Random
 using LineSearches
-using OptimizationBBO
 using ForwardDiff
 using LinearAlgebra
 using Statistics
