+++ src/plotting/plotting_residuals.jl
@@ -82,8 +82,8 @@ end
         return isfinite(x) ? x : missing
     end
     v isa AbstractVector &&
-        @warn "Vector-valued (multivariate) observations are not supported in scalar "*
-        "residual metrics and are reported as missing." maxlog=1
+        @warn "Vector-valued (multivariate) observations are not supported in scalar " *
+              "residual metrics and are reported as missing." maxlog=1
     return missing
 end
